Move wait_next_state into harness code
