+++ daemon-tests/src/lib.rs
@@ -331,6 +331,29 @@ macro_rules! deliver_event {
     }};
 }
 
+/// Waits until the CFDs for both maker and taker are in the given state.
+#[macro_export]
+macro_rules! wait_next_state {
+    ($id:expr, $maker:expr, $taker:expr, $maker_state:expr, $taker_state:expr) => {
+        let wait_until_taker = next_with($taker.cfd_feed(), one_cfd_with_state($taker_state));
+        let wait_until_maker = next_with($maker.cfd_feed(), one_cfd_with_state($maker_state));
+
+        let (taker_cfd, maker_cfd) = tokio::join!(wait_until_taker, wait_until_maker);
+        let taker_cfd = taker_cfd.unwrap();
+        let maker_cfd = maker_cfd.unwrap();
+
+        assert_eq!(
+            taker_cfd.order_id, maker_cfd.order_id,
+            "order id mismatch between maker and taker"
+        );
+        assert_eq!(taker_cfd.order_id, $id, "unexpected order id in the taker");
+        assert_eq!(maker_cfd.order_id, $id, "unexpected order id in the maker");
+    };
+    ($id:expr, $maker:expr, $taker:expr, $state:expr) => {
+        wait_next_state!($id, $maker, $taker, $state, $state)
+    };
+}
+
 pub fn dummy_price() -> Price {
     Price::new(dec!(50_000)).expect("to not fail")
 }
